Move functions from component to utils
